# openapi/product.json
{
  "components": {
    "schemas": {
...
      "FileReference": {
        "additionalProperties": false,
        "description": "An ordered navigation value to one ordinary workspace file.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "FileReference",
        "type": "object"
      },
...
      "OperationFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProductFailureCode"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "suggested_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Next Step"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "code",
          "summary",
          "retryable"
        ],
        "title": "OperationFailure",
        "type": "object"
      },
...
      "ProductFailureCode": {
        "description": "Closed, product-safe failure vocabulary shared by HTTP and Operator.",
        "enum": [
          "invalid_request",
          "not_found",
          "conflict",
          "cursor_reset_required",
          "access_denied",
          "unavailable",
          "internal_error"
        ],
        "title": "ProductFailureCode",
        "type": "string"
      },
...
      "TaskActivity": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskActivityLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FileReference"
            },
            "title": "Files",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TaskActivityKind"
          },
          "member": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskMemberReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskActivityOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "occurred_at",
          "title"
        ],
        "title": "TaskActivity",
        "type": "object"
      },
      "TaskActivityKind": {
        "enum": [
          "task_started",
          "task_paused",
          "task_resumed",
          "task_cancelled",
          "task_completed",
          "task_blocked",
          "work_completed",
          "work_blocked",
          "input_requested",
          "input_received",
          "input_expired",
          "input_cancelled",
          "action_started",
          "action_succeeded",
          "action_failed",
          "action_timed_out",
          "action_cancelled"
        ],
        "type": "string"
      },
      "TaskActivityLink": {
        "additionalProperties": false,
        "properties": {
          "href": {
            "title": "Href",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "href"
        ],
        "title": "TaskActivityLink",
        "type": "object"
      },
      "TaskActivityOutcome": {
        "enum": [
          "completed",
          "blocked",
          "failed",
          "cancelled"
        ],
        "type": "string"
      },
      "TaskActivityPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskActivity"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "TaskActivityPage",
        "type": "object"
      },
...
      "TaskMemberReference": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TaskMemberReference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Banksia Product API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tasks/{task_id}/activities": {
      "get": {
        "operationId": "get_task_activities_tasks__task_id__activities_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskActivityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Task Activities",
        "tags": [
          "task-activity"
        ]
      }
    },
    "/tasks/{task_id}/activities/stream": {
      "get": {
        "operationId": "stream_task_activities_tasks__task_id__activities_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationFailure"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Stream Task Activities",
        "tags": [
          "task-activity"
        ]
      }
    },
...
  }
}